requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pdf2image>=1.16.3
opencv-python>=4.8.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
